+++ src/components/Modal/Styles.js
@@ -1,0 +1,51 @@
+/** @format */
+
+import { makeStyles } from '@material-ui/core/styles';
+
+function getModalStyle() {
+  const top = 50;
+  const left = 50;
+
+  return {
+    top: `${top}%`,
+    left: `${left}%`,
+    transform: `translate(-${top}%, -${left}%)`,
+  };
+}
+
+const useStyles = makeStyles((theme) => ({
+  form: {
+    disaply: 'flex',
+    flexDirection: 'column',
+  },
+  paper: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    position: 'absolute',
+    width: 250,
+    backgroundColor: theme.palette.background.paper,
+    border: '2px solid #000',
+    boxShadow: theme.shadows[5],
+    padding: theme.spacing(2, 4, 3),
+  },
+  box: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'space-around',
+  },
+  button: {
+    marginBottom: 10,
+  },
+  logbutton: {
+    marginRight: 10,
+  },
+  img: {
+    maxWidth: 100,
+  },
+  login: {
+    dispaly: 'flex',
+  },
+}));
+
+export { getModalStyle, useStyles };
